Define ExaDecimal for pyodbc to resolve test
