+++ doc/changes/unreleased.md
@@ -21,6 +21,7 @@ This allows us to use the latest dependencies, which do not have open vulnerabil
   - Reinstated the ArgSignatureTest which ensures that all visit_XYZ() in `_sql.Compiler` subclasses have `**kw`
 - #626: Reinstated `sqlalchemy` tests:
   - `TrueDivTest.test_floordiv_integer` and `TrueDivTest.test_floordiv_integer_bound` by providing an override in `EXACompiler.visit_floordiv_binary`
+  - `TrueDivTest.test_truediv_numeric` by providing `ExaDecimal` to the `EXADialect_pyodbc.colspecs` list
 
 ## Internal
