+++ sqlalchemy_exasol/pyodbc.py
@@ -15,12 +15,14 @@ from packaging import version
 from sqlalchemy import sql
 from sqlalchemy.connectors.pyodbc import PyODBCConnector
 from sqlalchemy.engine import reflection
+from sqlalchemy.sql import sqltypes
 from sqlalchemy.util.langhelpers import asbool
 
 from sqlalchemy_exasol.base import (
     EXADialect,
     EXAExecutionContext,
 )
+from sqlalchemy_exasol.types import ExaDecimal
 from sqlalchemy_exasol.warnings import SqlaExasolDeprecationWarning
 
 logger = logging.getLogger("sqlalchemy_exasol")
@@ -30,6 +32,9 @@ class EXADialect_pyodbc(EXADialect, PyODBCConnector):
     supports_statement_cache = False
     execution_ctx_cls = EXAExecutionContext
     driver_version = None
+    colspecs = {
+        sqltypes.Numeric: ExaDecimal,
+    }
 
     def __init__(self, **kw):
         message = (
