+++ sqlalchemy_exasol/websocket.py
@@ -1,5 +1,4 @@
 import datetime
-import decimal
 import time
 from collections import (
     ChainMap,
@@ -11,6 +10,7 @@ from sqlalchemy.exc import ArgumentError
 from sqlalchemy.sql import sqltypes
 
 from sqlalchemy_exasol.base import EXADialect
+from sqlalchemy_exasol.types import ExaDecimal
 from sqlalchemy_exasol.version import VERSION
 
 
@@ -22,28 +22,6 @@ class Integer(sqltypes.INTEGER):
             return value
 
         return to_integer
-
-
-class Decimal(sqltypes.DECIMAL):
-    def bind_processor(self, dialect):
-        return super().bind_processor(dialect)
-
-    def result_processor(self, dialect, coltype):
-        if not self.asdecimal:
-            return lambda value: None if value is None else float(value)
-
-        fstring = "%%.%df" % self._effective_decimal_return_scale
-
-        def to_decimal(value):
-            if value is None:
-                return None
-            if isinstance(value, decimal.Decimal):
-                return value
-            if isinstance(value, float):
-                return decimal.Decimal(fstring % value)
-            return decimal.Decimal(value)
-
-        return to_decimal
 
 
 class Date(sqltypes.DATE):
@@ -89,7 +67,7 @@ class EXADialect_websocket(EXADialect):
     supports_statement_cache = False
     colspecs = {
         sqltypes.Integer: Integer,
-        sqltypes.Numeric: Decimal,
+        sqltypes.Numeric: ExaDecimal,
         sqltypes.Date: Date,
         sqltypes.DateTime: DateTime,
     }
